Tests/LibWeb/Text/expected/navigation/iframe-navigate-src.txt

message from test iframe (1)
message from test iframe (2)
